pyproject.toml
[tool.poetry]
name = "fastapi-codebase"
version = "0.0.1"
description = "FastAPI Backend Service Template"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.14.2"
fastapi = "0.128.0"
uvicorn = "0.40.0"
python-dotenv = "1.2.1"
sqlalchemy = "2.0.45"
pymysql = "1.1.2"
marshmallow = "4.2.0"
requests = "2.32.5"
python-dateutil = "2.9.0.post0"


[tool.poetry.group.dev.dependencies]
pytest = "9.0.2"
pytest-fastapi = "0.1.0"
ruff = "0.14.11"
ipython = "9.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "app"]
